requirements.txt
numpy
keras===2.12.0
tensorflow===2.12.0
opencv-python
Pillow
matplotlib